Define MapLayerConfig and ColumnInfo locally, remove geo-agent type re-exports
